Updated version helper to list both the main bundle version and the examples package version
