--- examples/ui.apps/src/content/jcr_root/apps/core-components-examples/components/version/CoreComponentsBundle.java
@@ -1,39 +1,0 @@
-/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
- ~ Copyright 2021 Adobe
- ~
- ~ Licensed under the Apache License, Version 2.0 (the "License");
- ~ you may not use this file except in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~     http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing, software
- ~ distributed under the License is distributed on an "AS IS" BASIS,
- ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- ~ See the License for the specific language governing permissions and
- ~ limitations under the License.
- ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-
-import org.osgi.framework.FrameworkUtil;
-
-import com.adobe.cq.sightly.WCMUsePojo;
-import com.adobe.cq.wcm.core.components.models.Page;
-
-/**
- * Helper Use-Object for displaying the Core Components bundle version
- *
- * Usage:
- *    <sly data-sly-use.bundle="CoreComponentsBundle">${bundle.version}</sly>
- */
-public class CoreComponentsBundle extends WCMUsePojo {
-
-    @Override
-    public void activate() throws Exception {
-
-    }
-
-    public String getVersion() {
-        return FrameworkUtil.getBundle(Page.class).getVersion().toString();
-    }
-
-}
