+++ examples/ui.apps/src/content/jcr_root/apps/core-components-examples/components/version/footer/footer.html
@@ -13,4 +13,4 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
-<p data-sly-use.bundle="CoreComponentsBundle">Built with <a href="https://github.com/adobe/aem-core-wcm-components/releases/">Core Components ${bundle.version}</a>.</p>
+<p data-sly-use.versionHelper="VersionHelper">Examples ${versionHelper.examplesVersion}, built with <a href="https://github.com/adobe/aem-core-wcm-components/releases/">Core Components ${versionHelper.bundleVersion}</a>.</p>
