+++ examples/ui.apps/src/content/jcr_root/apps/core-components-examples/components/version/title/title.html
@@ -13,8 +13,8 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
-<div class="cmp-examples-text--page-title" data-sly-use.bundle="CoreComponentsBundle">
+<div class="cmp-examples-text--page-title" data-sly-use.versionHelper="VersionHelper">
     <div class="cmp-text">
-        <h1>Component Library <a href="https://github.com/adobe/aem-core-wcm-components/releases/">${bundle.version}</a></h1>
+        <h1>Component Library <a href="https://github.com/adobe/aem-core-wcm-components/releases/">Bundle:${versionHelper.bundleVersion}, Examples: ${versionHelper.examplesVersion}</a></h1>
     </div>
 </div>
